+++ vast_pipeline/models.py
@@ -174,7 +174,7 @@ class SurveySource(models.Model):
 
 class RunQuerySet(models.QuerySet):
 
-    def check_max_runs(self, max_runs: int=5) -> int:
+    def check_max_runs(self, max_runs: int = 5) -> int:
         """
         Check if number of running pipeline runs is above threshold.
 
@@ -185,6 +185,9 @@ class RunQuerySet(models.QuerySet):
             The count of the current pipeline runs with a status of `RUN`.
         """
         return self.filter(status='RUN').count() >= max_runs
+
+
+RunManager = models.Manager.from_queryset(RunQuerySet)
 
 
 class Run(CommentableModel):
@@ -263,7 +266,7 @@ class Run(CommentableModel):
         )
     )
 
-    objects = RunQuerySet.as_manager()
+    objects = RunManager()  # used instead of RunQuerySet.as_manager() so mypy checks work
 
     class Meta:
         ordering = ['name']
@@ -445,33 +448,33 @@ class Source(CommentableModel):
     vs_abs_significant_max_int = models.FloatField(
         default=0.0,
         help_text=(
-            'Maximum value of all measurement pair variability t-statistics for int '
-            'flux that exceed SOURCE_AGGREGATE_PAIR_METRICS_MIN_ABS_VS in the pipeline run '
-            'configuration.'
+            'Maximum value of all measurement pair variability t-statistics for int'
+            ' flux that exceed variability.source_aggregate_pair_metrics_min_abs_vs in'
+            ' the pipeline run configuration.'
         )
     )
     m_abs_significant_max_int = models.FloatField(
         default=0.0,
         help_text=(
-            'Maximum absolute value of all measurement pair modulation indices for int '
-            'flux that exceed SOURCE_AGGREGATE_PAIR_METRICS_MIN_ABS_VS in the pipeline run '
-            'configuration.'
+            'Maximum absolute value of all measurement pair modulation indices for int'
+            ' flux that exceed variability.source_aggregate_pair_metrics_min_abs_vs in'
+            ' the pipeline run configuration.'
         )
     )
     vs_abs_significant_max_peak = models.FloatField(
         default=0.0,
         help_text=(
-            'Maximum absolute value of all measurement pair variability t-statistics for '
-            'peak flux that exceed SOURCE_AGGREGATE_PAIR_METRICS_MIN_ABS_VS in the pipeline '
-            'run configuration.'
+            'Maximum absolute value of all measurement pair variability t-statistics for'
+            ' peak flux that exceed variability.source_aggregate_pair_metrics_min_abs_vs'
+            ' in the pipeline run configuration.'
         )
     )
     m_abs_significant_max_peak = models.FloatField(
         default=0.0,
         help_text=(
             'Maximum absolute value of all measurement pair modulation indices for '
-            'peak flux that exceed SOURCE_AGGREGATE_PAIR_METRICS_MIN_ABS_VS in the '
-            'pipeline run configuration.'
+            ' peak flux that exceed variability.source_aggregate_pair_metrics_min_abs_vs'
+            ' in the pipeline run configuration.'
         )
     )